+++ src/service/evaluate/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "json")]
+pub mod json;
