+++ src/service/mod.rs
@@ -1,3 +1,4 @@
+pub mod evaluate;
 pub mod evaluator;
 pub mod factory;
